Add deps on mocha and requirejs (and alpha sort deps list)
In a fresh install, `npm install` outputs:

$ npm install
npm WARN karma-mocha@0.1.10 requires a peer of mocha@* but none was
installed.
npm WARN karma-requirejs@0.2.3 requires a peer of requirejs@~2.1 but
none was installed.

And sure enough the test fail without these packages.

This change adds mocha and requirejs as local development dependencies.