Remove initializer block for assets precompile
This block was commented out for testing. Its contents are not required
in a host app. Both Frontend and RAD repeat the lines related to
stylesheets in their own asset configuration.

Its unclear what affect removing the font regex will have in Frontend.
This will be tested as part of the process of merging these
changes into master.