Update README with usage instructions
Clarify how this should be added to the Gemfile of a host app.

Somewhere between Rails 4 and Rails 5, if the main file of the gem
doesn't match the gem name it now seems to be necessary to explicitly
state the name of the file so that the engine loads correctly. Make this
a bit more obvious in the README.